pyproject.toml
[tool.poetry]
name = "example-agents-project"
version = "0.1.0"
description = ""
authors = ["Thomas French <tommyfrench@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
httpx = "^0.28.1"
python-dotenv = "^1.0.1"
langgraph = "^0.2.61"
langchain-google-community = "^2.0.4"
openai = "^1.59.6"
langchain-openai = "^0.3.0"
psycopg2 = "^2.9.10"
asyncpg = "^0.30.0"
redis = "^5.2.1"
langgraph-checkpoint-postgres = "^2.0.10"


[tool.poetry.group.dev.dependencies]
ruff = "^0.9.0"
mypy = "^1.14.1"
pytest = "^8.3.4"
pytest-asyncio = "^0.25.2"
pytest-cov = "^6.0.0"
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "asyncio: mark a test as using asyncio"
]
